split code into 3 different classes
